requirements.txt
librosa==0.8.0
numpy==1.19.1
pandas==1.1.3
python-speech-features==0.6
PyYAML==5.4.1
scikit-learn==0.23.2
scipy==1.5.2
SoundFile==0.10.3
torch==1.2.0
torchsummary==1.5.1
